Assignment4/Q3/requirements.txt
torch
torchvision
transformers
clip-by-openai
pycocoevalcap
matplotlib
seaborn
scikit-learn
Pillow
gdown
opencv-python
numpy
nltk